+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### ⚠ Breaking
 - require Jenkins 2.550 or newer
 - require Java 21 or newer (required since Jenkins 2.545)
+- remove the "auto-create" option. it is now always "true". Users of [Configuration as Code](https://www.jenkins.io/projects/jcasc/) will have to remove the `dependencyTrackAutoCreateProjects` parameter.
 
 ### ⭐ New Features
 - Support uploading VEX files ([#271](https://github.com/jenkinsci/dependency-track-plugin/issues/271))